Fjernet bruk av lab2 og testl2, lagt til funksjonalitet for å se på hvilken sti programmet har tatt til åpningene den finner
