+++ Labyrint.java
@@ -1,58 +1,60 @@
 import java.util.Scanner;
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.util.ArrayList;
 
-
-public class Labyrint {
+public class Labyrint  {
     Rute[][] ruter;
     int antRader;
     int antKolonner;
     Scanner les;
     Scanner sc;
-    File innfil;
+    //File innfil;
+    ArrayList<ArrayList<Tuppel>> loesninger;
     
 
-    public Labyrint(File fil){
+    public Labyrint (String filbane){
         antRader = 0;
         antKolonner = 0;
-        innfil = fil;
-        lagLabyrint();
+        //innfil = fil;
+        lagLabyrint(filbane);
         kobleRuter();
+        loesninger = new ArrayList<ArrayList<Tuppel>>();
         }
     //Tar inn dimensjonene på labyrinten og oppretter en matrise
     //Leser filen linje for linje, og char for char for å opprette riktig ruteobjekt på hver index i matrisen
 
-    public void lagLabyrint(){
+    public void lagLabyrint(String filbane) {
         int i = 0;
-        try{
+        File innfil = new File(filbane);
+        try {
             sc = new Scanner(innfil);
-                antRader = sc.nextInt();
-                antKolonner = sc.nextInt();
-                ruter = new Rute[antRader][antKolonner];
-                sc.nextLine();
-                while (sc.hasNextLine()){
-                    String line = sc.nextLine();
-                    for(int j = 0; j <antKolonner; j++){            
-                        //Går igjennom hver char på linjen og sjekker hva slags rute som skal lages på bakgrunn av charen
-                        if(line.charAt(j) == '#'){                  
-                            ruter[i][j] = new SvartRute(i,j,this);
+            antRader = sc.nextInt();
+            antKolonner = sc.nextInt();
+            ruter = new Rute[antRader][antKolonner];
+            sc.nextLine();
+            while (sc.hasNextLine()){
+                String line = sc.nextLine();
+                for(int j = 0; j <antKolonner; j++){            
+                    //Går igjennom hver char på linjen og sjekker hva slags rute som skal lages på bakgrunn av charen
+                    if(line.charAt(j) == '#'){                  
+                        ruter[i][j] = new SvartRute(i,j,this);
 
-                        }else if(line.charAt(j) == '.'){                                        //sjekker om den aktuelle, hvite ruten er på siden til en av brettene, og oppretter
-                            if(i == 0 || i == antRader-1 || j == 0 || j == antKolonner -1){     //en instans av klassen Aapning om dette er tilfelle
-                                ruter[i][j] = new Aapning(i, j, this);
-                            }else{
-                            ruter[i][j] = new HvitRute(i,j,this);                               //Alle indre hvite ruter blir vanlige hvite ruter
-                            }
+                    }else if(line.charAt(j) == '.'){                                        //sjekker om den aktuelle, hvite ruten er på siden til en av brettene, og oppretter
+                        if(i == 0 || i == antRader-1 || j == 0 || j == antKolonner -1){     //en instans av klassen Aapning om dette er tilfelle
+                            ruter[i][j] = new Aapning(i, j, this);
+                        }else{
+                        ruter[i][j] = new HvitRute(i,j,this);                               //Alle indre hvite ruter blir vanlige hvite ruter
                         }
                     }
-                    i++;
-                        
                 }
-                }catch(FileNotFoundException e){
-                    System.out.println("fant ikke filen");
-                }finally{
-                    sc.close();
-                }
+                i++;
+            }
+            sc.close();
+            }
+            catch (FileNotFoundException e) {
+                System.err.println("fant ikke filen");
+            }
     }
     //Går igjennom matrisen og kaller på hver rutes "settnaboer()" metode, og kobler rutene sammen
     public void kobleRuter(){
@@ -66,13 +68,14 @@ public class Labyrint {
 
     public void finnUtveifra(int rad, int kolonne){
         if(ruter[rad][kolonne].hentFarge() == "hvit"){
-        ruter[rad][kolonne].finn(null);
+        ruter[rad][kolonne].finn(null, ruter[rad][kolonne].hentSti()); //Kaller på rute X Y sin finn() metode med fra = null && sti = X Y .hentSti()
+        System.out.println(loesninger);
         }else{
             System.out.println("Kan ikke starte paa sort rute");
         }
     }
-    public void tegnLab(){
 
+    public void tegnLab(){
         for(int i = 0; i<antRader;i++){
             for(int j= 0; j<antKolonner;j++){
                 System.out.print(ruter[i][j].hentTegn() + "  ");
@@ -100,7 +103,12 @@ public class Labyrint {
             System.out.println();
         }
     }
-
+    public int[] hentKordinater(Scanner sc){
+        int[] kordinater = new int[2];
+        kordinater[0] = sc.nextInt();
+        kordinater[1] = sc.nextInt();
+        return kordinater;
+    }
 
     @Override
     public String toString() {
