+++ Test.java
@@ -1,76 +1,71 @@
 import java.util.Scanner;
-import java.io.File;
+
 
 public class Test {
-    /*public static void main (String[] args) {
-        String input = "";
-        Scanner sc = new Scanner(System.in);
-        do {
-            
-            input = sc.nextLine();
-
-        }
-        while (!input.equals("-1"));
-        sc.close();
-
-        Labyrint2 lab = new Labyrint2("2.in");
-        System.out.println(lab);
-    }
-
-
-}*/
-
-    
-    protected static File fil;
-    protected static Labyrint lab;
-    static Scanner in;
     public static void main(String[] args) {
-        System.out.println("Skriv inn filnavnet paa labyrinten du vil se paa");
-        Scanner in = new Scanner(System.in);
-        velgBrett(in.nextLine(), in);
-        System.out.println("Tast 1 for aa spille, tast -1 for aa avslutte, tast 2 for aa velge nytt brett");
+        Scanner sc = new Scanner(System.in);
+        String userInp = "";
+        Labyrint lab = null;
+        int[] kordinater;
         
-        velg(in);
-}
-    private static void velgBrett(String arg, Scanner in ){
-        String filnavn = arg;
-        fil = new File(filnavn);
-        Labyrint lab = new Labyrint(fil);
-        System.out.println("Slik ser brettet ut");
-        lab.tegnLab();
-    }
+        System.out.println("Velkommen til labyrintløseren");
 
-    private static void spill(Scanner in){
-        int les = 0;
-        while (les != -1){
-        System.out.println("Velgstartspunktkordinater <rad> <kolonne>");
-        //Scanner in = new Scanner(System.in);
-        int rad = in.nextInt();
-        int kolonne = in.nextInt();
-        lab = new Labyrint(fil);
-        lab.finnUtveifra(rad, kolonne);
-        System.out.println("Tast 1 for aa spille, tast -1 for aa avslutte, eller 2 for aa bytte brett");
-        velg(in);
-    }
-    System.out.println("Spillet avslutter");
+        while (!userInp.equals("-1")) {
 
-    }
+            System.out.println("'1' for å laste inn ny labyrint\n'2' for å printe labyrint\n'3' for å vise åpninger\n'4' for å vise traversering\n'-1' for å avslutte");
+            userInp = sc.nextLine();
+            switch (userInp) {
+                case "1":
+                    lab = lesFraFil(sc);
+                    break;
 
-    private static void velg(Scanner in){
-        //Scanner in = new Scanner(System.in);
-        int les = in.nextInt();
-        while(les != -1){
-        if(les == 1){
-            spill(in);
-        } else if (les == 2){
-            System.out.println("Skriv inn filnavnet paa labyrinten du vil se paa: ");
-            velgBrett(in.nextLine(), in);
-            velg(in);
+                case "2":
+                    try {   
+                        lab.tegnLab();
+                    } catch (NullPointerException e) {
+                        System.err.println("Labyrinten er ikke lastet inn, forsøk å lese ny labyrint fra fil først");
+                    }
+                    break;
+
+                case "3":
+                try {
+                    System.out.println("Skriv inn kordinater på form <rad> <kolonne>");
+                    kordinater = lab.hentKordinater(sc);
+                    int rad = kordinater[0];
+                    int kolonne = kordinater[1];
+                    lab.finnUtveifra(rad, kolonne);
+                } catch (NullPointerException e) {
+                    System.err.println("Labyrinten er ikke lastet inn, forsøk å lese ny labyrint fra fil først");
+                }
+                    break;
+
+                case "4":
+                try {
+                    lab.tegnTraversering();
+                } catch (NullPointerException e) {
+                    System.err.println("Labyrinten er ikke lastet inn, forsøk å lese ny labyrint fra fil først");
+                }
+                    break;
+            }
         }
-        
+
+        sc.close();
     }
-        System.out.println("spillet avslutter");
-        in.close();
-        return;
+
+    public static Labyrint lesFraFil(Scanner scanner) {
+        System.out.println("Oppgi filbane eller skriv '-1' for å gå tilbake");
+        String inp = scanner.nextLine();
+        while (!inp.equals("-1")) {
+
+            Labyrint nyLab = new Labyrint(inp);
+            if (nyLab.antRader > 0) {
+                System.out.println("Lastet inn labyrint fra " + inp);
+                return nyLab;
+            } else {
+                inp = scanner.nextLine();
+            }
+        }
+        return null;
     }
+    
 }
